requirements.txt
dspy-ai>=2.5.0
pydantic>=2.0.0
pillow>=10.0.0
loguru>=0.7.0
omegaconf>=2.3.0
asyncio>=3.4.3
numpy>=1.24.0
python-dotenv>=1.0.0
httpx>=0.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
ruff>=0.1.0
mypy>=1.0.0
types-Pillow>=10.0.0
mkdocs>=1.5.0
mkdocs-material>=9.0.0
ipython>=8.0.0
jupyter>=1.0.0